requirements.txt
bert-for-tf2==0.14.4
glob2==0.7
joblib==0.16.0
matplotlib==3.0.0
numpy==1.19.0
pandas==1.0.5
scikit-learn==0.23.1
scikit-optimize==0.7.4
scipy==1.4.1
tensorflow-hub==0.8.0
tensorflow==2.2.0
tqdm==4.48.0
umap-learn==0.4.6
